transfer main to distinct class, add some tests, minor changes
